+++ src/nodejs/lambda-layers/database-util/src/query/submission.js
@@ -4,10 +4,14 @@ const workflow = require('./workflow.js');
 
 const table = 'submission';
 // const allFields = ['id', 'name', 'user_id', 'daac_id', 'conversation_id', 'workflow_id', 'workflow_name', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock'];
+
 const allFields = ['id', 'name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'step_data', 'step_status_label', 'step_name', 'status', 'forms', 'action_data', 'form_data', 'metadata', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
+const customFields = ['id', 'name', 'data_producer_name', 'initiator', 'workflow_id', 'hidden', 'conversation_id', 'workflow_name', 'daac_id', 'daac_name', 'step_data', 'step_name', 'status', 'created_at', 'last_change', 'lock', 'contributor_ids', 'copy', 'origin_id'];
+
 const fieldMap = {
   id: 'submission.id',
   name: 'submission.name',
+  data_producer_name: 'submission.data_producer_name',
   initiator: 'initiator_ref.initiator',
   user_id: 'submission.initiator_edpuser_id user_id',
   daac_id: 'submission.daac_id',
@@ -176,10 +180,10 @@ const findById = (params) => sql.select({
 });
 
 const getUsersSubmissions = (params) => sql.select({
-  fields: fields(allFields),
+  fields: fields(customFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
+    joins: [refs.submission_status, refs.initiator_ref, refs.step, refs.workflow, refs.daac, refs.submission_copy]
   },
   where: {
     filters: [
@@ -195,10 +199,10 @@ const getDaacSubmissions = (params) => sql.select({
   fields: ['*'],
   from: {
       base: `(${sql.select({
-        fields: fields(allFields),
+        fields: fields(customFields),
     from: {
     base: table,
-      joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
+      joins: [refs.submission_status, refs.initiator_ref, refs.step, refs.workflow, refs.daac, refs.submission_copy]
     },
         where: {
           filters: [
@@ -254,10 +258,10 @@ const getDaacSubmissions = (params) => sql.select({
 });
 
 const getAdminSubmissions = (params) => sql.select({
-  fields: fields(allFields),
+  fields: fields(customFields),
   from: {
     base: table,
-    joins: [refs.submission_status, refs.initiator_ref, refs.submission_metadata, refs.submission_action_data, refs.submission_form_data, refs.step, refs.workflow, refs.daac, refs.submission_form_data_pool, refs.submission_copy]
+    joins: [refs.submission_status, refs.initiator_ref, refs.step, refs.workflow, refs.daac, refs.submission_copy]
   },
   where: {
     filters: [
@@ -366,6 +370,13 @@ INSERT INTO submission_form_data(id, form_id, data) VALUES
 ON CONFLICT (id, form_id) DO UPDATE SET
 data = EXCLUDED.data;
 END $$`;
+
+const updateSubmissionData = () => `
+update submission 
+set name = {{data_product}}, data_producer_name = {{data_producer}}
+where id = {{id}}
+RETURNING *
+`;
 
 const getActionData = () => `
 SELECT
@@ -574,6 +585,7 @@ SELECT step_name FROM submission_status WHERE id = {{id}}
 // TODO - Upate this query's complexity and to use sql builder
 const getSubmissionDetailsById = () => `
 SELECT submission.id id, conversation_id, submission.created_at created_at, daac.long_name daac_name, 
+submission.hidden hidden,
 JSONB_BUILD_OBJECT('name', edpuser1.name, 'id', edpuser1.id) initiator,
 submission_status.last_change last_change,
 submission_form_data_pool.data::json->'data_producer_info_name' data_producer_name, 
@@ -633,6 +645,73 @@ const getSubmissionDaac = () => sql.select({
   }
 })
 
+
+const getStepReviewApproval = () => `
+SELECT sr.step_name, sr.submission_id, sr.edpuser_id, sr.user_review_status, eu.name, sr.submitted_by
+FROM step_review sr
+INNER JOIN edpuser eu ON sr.edpuser_id = eu.id
+WHERE sr.submission_id = {{id}}
+`;
+
+
+const createStepReviewApproval = (params) => `
+INSERT INTO step_review (step_name, submission_id, edpuser_id, user_review_status, submitted_by)
+SELECT {{step_name}}, {{submission_id}}, CAST(user_id AS UUID), 'review_required', {{submitted_by}}
+FROM unnest(ARRAY[${params.user_ids.map(id => `'${id}'::UUID`).join(',')}]) AS user_id
+RETURNING *
+`;
+
+
+const deleteStepReviewApproval = (params) => `
+DELETE FROM step_review
+WHERE submission_id = {{submission_id}}
+AND step_name = {{step_name}}
+AND edpuser_id IN (${params.user_ids.map(id => `'${id}'::UUID`).join(', ')})
+RETURNING *;
+`;
+
+const checkCountStepReviewApproved = () => `
+SELECT COALESCE(SUM(sr.unapproved_count), 0) AS unapproved
+FROM (
+    SELECT submission_id, step_name, 
+           SUM(CASE WHEN user_review_status IN ('rejected', 'review_required') THEN 1 ELSE 0 END) AS unapproved_count
+    FROM step_review
+    WHERE submission_id = {{submission_id}}
+      AND step_name = {{step_name}}
+    GROUP BY submission_id, step_name
+) AS sr
+LEFT JOIN (SELECT 1) AS dummy ON 1 = 1
+`;
+
+const checkCountStepReviewRejected = () => `
+SELECT COUNT(*) AS unapproved
+FROM step_review
+WHERE submission_id = {{submission_id}}
+  AND step_name = {{step_name}}
+  AND edpuser_id <> {{user_id}}
+`;
+
+
+const updateStatusStepReviewApproval = () => `
+UPDATE step_review
+SET user_review_status = {{approve}}
+WHERE submission_id = {{submission_id}}
+  AND edpuser_id = {{user_id}}
+  AND step_name = {{step_name}}
+RETURNING *
+`;
+
+const stepCleanup = () => `
+WITH step_to_delete AS (
+    SELECT s.step_name
+    FROM Step s
+    LEFT JOIN Step_Edge se1 ON s.step_name = se1.step_name
+    LEFT JOIN Step_Edge se2 ON s.step_name = se2.next_step_name
+    WHERE se1.step_name IS NULL AND se2.next_step_name IS NULL
+)
+DELETE FROM Step
+WHERE step_name IN (SELECT step_name FROM step_to_delete)`;
+
 module.exports.findAll = findAll;
 module.exports.findShortById = findShortById;
 module.exports.findById = findById;
@@ -653,6 +732,7 @@ module.exports.getActionData = getActionData;
 module.exports.updateActionData = updateActionData;
 module.exports.getFormData = getFormData;
 module.exports.updateFormData = updateFormData;
+module.exports.updateSubmissionData = updateSubmissionData;
 module.exports.applyWorkflow = applyWorkflow;
 module.exports.rollback = rollback;
 module.exports.reassignWorkflow = reassignWorkflow;
@@ -671,3 +751,10 @@ module.exports.getCreatorName = getCreatorName;
 module.exports.getStepName = getStepName;
 module.exports.getSubmissionDetailsById = getSubmissionDetailsById;
 module.exports.getSubmissionDaac = getSubmissionDaac;
+module.exports.getStepReviewApproval = getStepReviewApproval;
+module.exports.createStepReviewApproval = createStepReviewApproval;
+module.exports.deleteStepReviewApproval = deleteStepReviewApproval;
+module.exports.checkCountStepReviewRejected = checkCountStepReviewRejected;
+module.exports.checkCountStepReviewApproved = checkCountStepReviewApproved;
+module.exports.updateStatusStepReviewApproval = updateStatusStepReviewApproval;
+module.exports.stepCleanup = stepCleanup;