+++ src/nodejs/lambda-layers/database-util/src/query/user.js
@@ -75,10 +75,20 @@ const refs = {
     src: group.userJoin,
     on: { left: fieldMap.id, right: 'group_agg.edpuser_id' }
   },
+  simple_group: {
+    type: 'left_join',
+    src: 'edpuser_edpgroup',
+    on: { left: fieldMap.id, right: 'edpuser_edpgroup.edpuser_id' }
+  },
   role: {
     type: 'left_join',
     src: role.userJoin,
     on: { left: fieldMap.id, right: 'role_agg.edpuser_id' }
+  },
+  simple_role: {
+    type: 'left_join',
+    src: 'edpuser_edprole',
+    on: { left: fieldMap.id, right: 'edpuser_edprole.edpuser_id' }
   },
   user_permission_submission: {
     type: 'left_join',
@@ -220,31 +230,27 @@ const find = ({ id, name, email, sort, order, per_page, page }) => sql.select({
   ...(page ? { offset: page } : {})
 });
 
-const findAll = ({name, email, sort, order, per_page, page}) => sql.select({
-    fields: fields(allFields),
-    from: {
-      base: table,
-      joins: [
-        refs.group,
-        refs.role,
-        refs.user_permission_submission,
-        refs.user_subscription_action,
-        refs.user_subscription_form,
-        refs.user_subscription_service,
-        refs.user_subscription_submission,
-        refs.user_subscription_workflow
-      ]
-    },
-    where: {
-      filters: [
-          ...(name ? [{ field: 'edpuser.name', like: 'name' }] : []),
-          ...(email ? [{ field: 'edpuser.email', like: 'email'}] : [])
-      ]
-    },
-    ...(sort ? { sort } : {}),
-    ...(order ? { order } : {}),
-    ...(per_page ? { limit: per_page } : {}),
-    ...(page ? { offset: page } : {})
+const findAll = ({name, email, sort, order, per_page, page, group_id, role_id}) => sql.select({
+  fields: fields(['id', 'name']),
+  from: {
+    base: table,
+    joins: [
+      refs.simple_group,
+      refs.simple_role
+    ]
+  },
+  where: {
+    filters: [
+      ...(name ? [{ field: 'edpuser.name', like: 'name' }] : []),
+      ...(email ? [{ field: 'edpuser.email', like: 'email'}] : []),
+      ...(group_id ? [{ field: 'edpuser_edpgroup.edpgroup_id', param: 'group_id' }] : []),
+      ...(role_id ? [{ field: 'edpuser_edprole.edprole_id', param: 'role_id' }] : []),
+    ]
+  },
+  ...(sort ? { sort } : {}),
+  ...(order ? { order } : {}),
+  ...(per_page ? { limit: per_page } : {}),
+  ...(page ? { offset: page } : {})
 });
 
 const findById = () => sql.select({
@@ -400,7 +406,7 @@ name = {{name}}
 WHERE edpuser.id = {{id}}
 RETURNING *`;
 
-const getUnknownStaffIds = (params) => sql.select({
+const getStaffIds = (params) => sql.select({
   fields: ['id'],
   from: {
     base: 'edpuser',
@@ -411,8 +417,8 @@ const getUnknownStaffIds = (params) => sql.select({
   },
   where: {
     filters: [
-      { field: 'edpuser_edpgroup.edpgroup_id', literal: '5be24b44-d66b-4396-9266-a9d066000d9e' },
-      { field: 'edpuser_edprole.edprole_id', literal: '2aa89c57-85f1-4611-812d-b6760bb6295c' }
+      { cmd: "edpuser_edpgroup.edpgroup_id = (SELECT edpgroup_id FROM daac WHERE id = {{daac_id}})" },
+      { cmd: "(edpuser_edprole.edprole_id = 'a5b4947a-67d2-434e-9889-59c2fad39676' OR edpuser_edprole.edprole_id = '2aa89c57-85f1-4611-812d-b6760bb6295c')"}
     ]
   }
 })
@@ -438,4 +444,4 @@ module.exports.findByEmail = findByEmail;
 module.exports.getUsers = getUsers;
 module.exports.setDetail = setDetail;
 module.exports.updateUsername = updateUsername;
-module.exports.getUnknownStaffIds = getUnknownStaffIds;
+module.exports.getStaffIds = getStaffIds;