+++ terraform/apigateway/variables.tf
@@ -97,3 +97,7 @@ variable "region" {
 variable "service_authorizer_lambda_arn" {
   type = string
 }
+
+variable "mfa_auth_lambda_arn" {
+  type = string
+}