+++ terraform/aws_secrets/main.tf
@@ -11,3 +11,31 @@ resource "aws_secretsmanager_secret_version" "ses_access_creds" {
         "ses_secret_access_key" = var.ses_secret_access_key
     })
 }
+
+#ORNL enpoint access
+resource "aws_secretsmanager_secret" "ornl_endpoint" {
+  name = "ornl_endpoint"
+  description = "ORNL endpoint and credentials"
+}
+
+resource "aws_secretsmanager_secret_version" "ornl_endpoint" {
+  secret_id = aws_secretsmanager_secret.ornl_endpoint.id
+    secret_string = jsonencode({
+      "ornl_endpoint_url" = var.ornl_endpoint_url
+      "ornl_endpoint_access_token" = var.ornl_endpoint_access_token
+    })
+}
+
+#GESDISC endpoint access
+resource "aws_secretsmanager_secret" "gesdisc_endpoint"{
+  name = "gesdisc_endpoint"
+  description = "GESDISC endpoint and credentials"
+}
+
+resource "aws_secretsmanager_secret_version" "gesdisc_endpoint"{
+  secret_id = aws_secretsmanager_secret.gesdisc_endpoint.id
+    secret_string = jsonencode({
+      "gesdisc_endpoint_url" = var.gesdisc_endpoint_url
+      "gesdisc_endpoint_access_token" = var.gesdisc_endpoint_access_token
+    })
+}