+++ terraform/aws_secrets/variables.tf
@@ -5,3 +5,19 @@ variable "ses_access_key_id" {
 variable "ses_secret_access_key" {
   type = string
 }
+
+variable "ornl_endpoint_url" {
+  type = string
+}
+
+variable "ornl_endpoint_access_token" {
+  type = string
+}
+
+variable "gesdisc_endpoint_url"{
+  type = string
+}
+
+variable "gesdisc_endpoint_access_token"{
+  type = string
+}