+++ terraform/modules.tf
@@ -16,6 +16,8 @@ module "iam_roles" {
   lambda_execution_policy_arn = var.lambda_execution_policy_arn
   permissions_boundary_arn = var.permissions_boundary_arn
   ses_access_creds_arn = module.aws_secrets.ses_access_creds_arn
+  ornl_endpoint_arn = module.aws_secrets.ornl_endpoint_arn
+  gesdisc_endpoint_arn = module.aws_secrets.gesdisc_endpoint_arn
 }
 
 module "s3" {
@@ -29,6 +31,10 @@ module "aws_secrets" {
 
   ses_access_key_id = var.ses_access_key_id
   ses_secret_access_key = var.ses_secret_access_key
+  ornl_endpoint_url = var.ornl_endpoint_url
+  ornl_endpoint_access_token = var.ornl_endpoint_access_token
+  gesdisc_endpoint_url = var.gesdisc_endpoint_url
+  gesdisc_endpoint_access_token = var.gesdisc_endpoint_access_token
 }
 
 module "lambda_functions" {
@@ -81,6 +87,8 @@ module "lambda_functions" {
   meditor_service_password = var.meditor_service_password
   ses_from_email = var.ses_from_email
   ses_access_creds_arn = module.aws_secrets.ses_access_creds_arn
+  ornl_endpoint_arn = module.aws_secrets.ornl_endpoint_arn
+  gesdisc_endpoint_arn = module.aws_secrets.gesdisc_endpoint_arn
 }
 
 module "apigateway_endpoints" {
@@ -111,6 +119,7 @@ module "apigateway_endpoints" {
   vpc_endpoint_id = var.vpc_endpoint_id
   region = var.region
   service_authorizer_lambda_arn = module.lambda_functions.service_authorizer_lambda_arn
+  mfa_auth_lambda_arn = module.lambda_functions.mfa_auth_lambda_arn
 }
 
 module "rds" {