+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 <!-- Unreleased changes can be added here. -->
+- Update data_poc users to producer and remove data_poc and it's references
+- Added GES DISC extended questions
+- Created note_scope table
 - Updated metadata mapper for DAAC specific questions
 - Updated email notification text formatting to accommodate special characters
 - Updated contact information label to full name from first and last name
@@ -16,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Creates example group and daac.
 - Updates step promotion to update set event type programmatically.
 - Update cloud metrics logging
+- Reimplemented custom MFA flow
+- Added filter by group and role id to the users endpont
 
 ## 1.0.20